Refactored, parameterized tests
