+++ atlasdb-cassandra/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraRequestExceptionHandlerTest.java
@@ -16,11 +16,13 @@
 
 package com.palantir.atlasdb.keyvalue.cassandra;
 
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 
 import java.net.SocketTimeoutException;
+import java.util.Arrays;
 import java.util.HashSet;
 import java.util.NoSuchElementException;
 import java.util.Set;
@@ -29,41 +31,55 @@ import org.apache.cassandra.thrift.InvalidRequestException;
 import org.apache.cassandra.thrift.TimedOutException;
 import org.apache.cassandra.thrift.UnavailableException;
 import org.apache.thrift.transport.TTransportException;
+import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
 import com.palantir.atlasdb.cassandra.CassandraKeyValueServiceConfig;
 import com.palantir.atlasdb.keyvalue.api.InsufficientConsistencyException;
 
+@RunWith(Parameterized.class)
 public class CassandraRequestExceptionHandlerTest {
-
     private static final String MESSAGE = "a exception";
     private static final Exception CAUSE = new Exception();
     private static final CassandraKeyValueServiceConfig config = mock(CassandraKeyValueServiceConfig.class);
     private static final int MAX_RETRIES_PER_HOST = 3;
     private static final int MAX_RETRIES_TOTAL = 6;
 
-    private Set<Exception> connectionExceptions = Sets.newHashSet(new SocketTimeoutException(MESSAGE),
+    private static Set<Exception> connectionExceptions = Sets.newHashSet(new SocketTimeoutException(MESSAGE),
             new CassandraClientFactory.ClientCreationFailedException(MESSAGE, CAUSE));
-    private Set<Exception> transientExceptions = Sets.newHashSet(new TTransportException(),
+    private static Set<Exception> transientExceptions = Sets.newHashSet(new TTransportException(),
             new InsufficientConsistencyException(MESSAGE));
-    private Set<Exception> indicativeOfCassandraLoadException = Sets.newHashSet(new NoSuchElementException(),
+    private static Set<Exception> indicativeOfCassandraLoadException = Sets.newHashSet(new NoSuchElementException(),
             new TimedOutException(), new UnavailableException());
-    private Set<Exception> fastFailoverExceptions = Sets.newHashSet(new InvalidRequestException());
+    private static Set<Exception> fastFailoverExceptions = Sets.newHashSet(new InvalidRequestException());
 
-    private DefaultRequestExceptionHandler defaultHandler = new DefaultRequestExceptionHandler(
-            () -> MAX_RETRIES_PER_HOST,
-            () -> MAX_RETRIES_TOTAL,
-            new Blacklist(config));
+    private boolean currentMode;
+    private CassandraRequestExceptionHandler handler;
 
-    private ConservativeRequestExceptionHandler conservativeHandler = new ConservativeRequestExceptionHandler(
-            () -> MAX_RETRIES_PER_HOST,
-            () -> MAX_RETRIES_TOTAL,
-            new Blacklist(config));
+    @Parameterized.Parameters
+    public static Iterable<? extends Object> data() {
+        return Arrays.asList(false, true);
+    }
+
+    @Parameterized.Parameter
+    public boolean useConservative;
+
+    @Before
+    public void setup() {
+        currentMode = useConservative;
+        handler  = new CassandraRequestExceptionHandler(
+                () -> MAX_RETRIES_PER_HOST,
+                () -> MAX_RETRIES_TOTAL,
+                this::mutableMode,
+                new Blacklist(config));
+    }
 
     @Test
-    public void testShouldRetryTest() {
+    public void retryableExceptionsAreRetryable() {
         Set<Exception> allExceptions = new HashSet<>();
         allExceptions = Sets.union(allExceptions, connectionExceptions);
         allExceptions = Sets.union(allExceptions, transientExceptions);
@@ -71,74 +87,53 @@ public class CassandraRequestExceptionHandlerTest {
         allExceptions = Sets.union(allExceptions, fastFailoverExceptions);
 
         for (Exception ex : allExceptions) {
-            assertTrue(String.format("Exception %s should be retryable", ex), defaultHandler.isRetryable(ex));
-            assertTrue(String.format("Exception %s should be retryable", ex), conservativeHandler.isRetryable(ex));
+            assertTrue(String.format("Exception %s should be retryable", ex), handler.isRetryable(ex));
         }
-
-        assertFalse("RuntimeException is not retryable", defaultHandler.isRetryable(new RuntimeException()));
-        assertFalse("RuntimeException is not retryable", conservativeHandler.isRetryable(new RuntimeException()));
+        assertFalse("RuntimeException is not retryable", handler.isRetryable(new RuntimeException()));
     }
 
     @Test
-    public void shouldBlacklistTest() {
+    public void connectionExceptionsWithSufficientAttemptsShouldBlacklist() {
         for (Exception ex : connectionExceptions) {
             assertFalse("MAX_RETRIES_PER_HOST - 1 attempts should not blacklist",
-                    defaultHandler.shouldBlacklist(ex, MAX_RETRIES_PER_HOST - 1));
-            assertFalse("MAX_RETRIES_PER_HOST - 1 attempts should not blacklist",
-                    conservativeHandler.shouldBlacklist(ex, MAX_RETRIES_PER_HOST - 1));
+                    handler.shouldBlacklist(ex, MAX_RETRIES_PER_HOST - 1));
         }
 
         for (Exception ex : connectionExceptions) {
             assertTrue(String.format("MAX_RETRIES_PER_HOST attempts with exception %s should blacklist", ex),
-                    defaultHandler.shouldBlacklist(ex, MAX_RETRIES_PER_HOST));
-            assertTrue(String.format("MAX_RETRIES_PER_HOST attempts with exception %s should blacklist", ex),
-                    conservativeHandler.shouldBlacklist(ex, MAX_RETRIES_PER_HOST));
+                    handler.shouldBlacklist(ex, MAX_RETRIES_PER_HOST));
         }
 
         Exception ffException = Iterables.get(fastFailoverExceptions, 0);
         assertFalse(String.format("Exception %s should not blacklist", ffException),
-                defaultHandler.shouldBlacklist(ffException, MAX_RETRIES_PER_HOST));
-        assertFalse(String.format("Exception %s should not blacklist", ffException),
-                conservativeHandler.shouldBlacklist(ffException, MAX_RETRIES_PER_HOST));
+                handler.shouldBlacklist(ffException, MAX_RETRIES_PER_HOST));
     }
 
     @Test
-    public void connectionExceptionsShouldBackoffTest() {
+    public void connectionExceptionsShouldBackoff() {
         for (Exception ex : connectionExceptions) {
-            assertTrue(String.format("Exception %s should backoff", ex),
-                    defaultHandler.shouldBackoff(ex));
-            assertTrue(String.format("Exception %s should backoff", ex),
-                    conservativeHandler.shouldBackoff(ex));
+            assertTrue(String.format("Exception %s should backoff", ex), handler.shouldBackoff(ex));
         }
     }
 
     @Test
-    public void cassandraLoadExceptionsShouldBackoffTest() {
+    public void cassandraLoadExceptionsShouldBackoff() {
         for (Exception ex : indicativeOfCassandraLoadException) {
-            assertTrue(String.format("Exception %s should backoff", ex),
-                    defaultHandler.shouldBackoff(ex));
-            assertTrue(String.format("Exception %s should backoff", ex),
-                    conservativeHandler.shouldBackoff(ex));
+            assertTrue(String.format("Exception %s should backoff", ex), handler.shouldBackoff(ex));
         }
     }
 
     @Test
     public void transientExceptionsShouldBackoffForConservativeOnly() {
         for (Exception ex : transientExceptions) {
-            assertFalse(String.format("Exception %s should not backoff", ex),
-                    defaultHandler.shouldBackoff(ex));
-            assertTrue(String.format("Exception %s should backoff", ex),
-                    conservativeHandler.shouldBackoff(ex));
+            assertEquals(useConservative, handler.shouldBackoff(ex));
         }
     }
 
     @Test
     public void fastFailoverExceptionsShouldNotBackoff() {
         for (Exception ex : fastFailoverExceptions) {
-            assertFalse(String.format("Exception %s should not backoff", ex),
-                    defaultHandler.shouldBackoff(ex));
-            assertFalse(String.format("Exception %s should not backoff", ex),
-                    conservativeHandler.shouldBackoff(ex));
+            assertFalse(String.format("Exception %s should not backoff", ex), handler.shouldBackoff(ex));
         }
     }
 
@@ -146,42 +141,27 @@ public class CassandraRequestExceptionHandlerTest {
     public void connectionExceptionRetriesOnDifferentHostAfterSufficientRetries() {
         for (Exception ex : connectionExceptions) {
             assertFalse(String.format("Exception %s should not retry on different host", ex),
-                    defaultHandler.shouldRetryOnDifferentHost(ex, MAX_RETRIES_PER_HOST - 1));
-            assertFalse(String.format("Exception %s should not retry on different host", ex),
-                    conservativeHandler.shouldRetryOnDifferentHost(ex, MAX_RETRIES_PER_HOST - 1));
+                    handler.shouldRetryOnDifferentHost(ex, MAX_RETRIES_PER_HOST - 1));
         }
 
         for (Exception ex : connectionExceptions) {
             assertTrue(String.format("Exception %s should retry on different host", ex),
-                    defaultHandler.shouldRetryOnDifferentHost(ex, MAX_RETRIES_PER_HOST));
-            assertTrue(String.format("Exception %s should retry on different host", ex),
-                    conservativeHandler.shouldRetryOnDifferentHost(ex, MAX_RETRIES_PER_HOST));
+                    handler.shouldRetryOnDifferentHost(ex, MAX_RETRIES_PER_HOST));
         }
     }
 
     @Test
-    public void cassandraLoadExceptionRetriesOnDifferentHostAfterSufficientRetriesForDefaultHandler() {
+    public void cassandraLoadExceptionRetriesOnDifferentHostWithoutRetryingForConservativeHandlerOnly() {
         for (Exception ex : indicativeOfCassandraLoadException) {
-            assertFalse(String.format("Exception %s should not retry on different host", ex),
-                    defaultHandler.shouldRetryOnDifferentHost(ex, MAX_RETRIES_PER_HOST - 1));
-        }
-
-        for (Exception ex : indicativeOfCassandraLoadException) {
-            assertTrue(String.format("Exception %s should retry on different host", ex),
-                    defaultHandler.shouldRetryOnDifferentHost(ex, MAX_RETRIES_PER_HOST));
+            assertEquals(useConservative, handler.shouldRetryOnDifferentHost(ex, 0));
         }
     }
 
     @Test
-    public void cassandraLoadExceptionAlwaysRetriesOnDifferentHostForConservativeHandler() {
+    public void cassandraLoadExceptionRetriesOnDifferentHostAfterSufficientRetries() {
         for (Exception ex : indicativeOfCassandraLoadException) {
             assertTrue(String.format("Exception %s should retry on different host", ex),
-                    conservativeHandler.shouldRetryOnDifferentHost(ex, MAX_RETRIES_PER_HOST - 1));
-        }
-
-        for (Exception ex : indicativeOfCassandraLoadException) {
-            assertTrue(String.format("Exception %s should retry on different host", ex),
-                    defaultHandler.shouldRetryOnDifferentHost(ex, MAX_RETRIES_PER_HOST));
+                    handler.shouldRetryOnDifferentHost(ex, MAX_RETRIES_PER_HOST));
         }
     }
 
@@ -189,8 +169,27 @@ public class CassandraRequestExceptionHandlerTest {
     public void fastFailoverExceptionAlwaysRetriesOnDifferentHost() {
         for (Exception ex : fastFailoverExceptions) {
             assertTrue(String.format("Fast failover exception %s should always retry on different host", ex),
-                    defaultHandler.shouldRetryOnDifferentHost(ex, 0));
-            assertTrue(String.format("Fast failover exception %s should always retry on different host", ex),
-                    conservativeHandler.shouldRetryOnDifferentHost(ex, 0));        }
+                    handler.shouldRetryOnDifferentHost(ex, 0));
+        }
+    }
+
+    @Test
+    public void changingHandlerModeHasNoEffectUntilUpdateStrategy() {
+        Exception ex = Iterables.get(transientExceptions, 0);
+        assertEquals(useConservative, handler.shouldBackoff(ex));
+
+        flipMode();
+        assertEquals(useConservative, handler.shouldBackoff(ex));
+
+        handler.updateStrategy();
+        assertEquals(!useConservative, handler.shouldBackoff(ex));
+    }
+
+    private boolean mutableMode() {
+        return currentMode;
+    }
+
+    private void flipMode() {
+        currentMode = !currentMode;
     }
 }
